# roller-network/src/main/resources/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "High Roller's Network Management",
      "description": "This Rest API provides the services to manage the online casino and its network of players."
    }
  ],
  "paths": {
    "/highrollernetwork/player/{name}/transfer": {
      "put": {
        "tags": [
          "High Roller's Network Management"
        ],
        "operationId": "playerTransfer",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 50,
              "minLength": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/highrollernetwork/player/{name}/exit": {
      "put": {
        "tags": [
          "High Roller's Network Management"
        ],
        "operationId": "playerExit",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 50,
              "minLength": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/highrollernetwork/player": {
      "get": {
        "tags": [
          "High Roller's Network Management"
        ],
        "operationId": "getPlayerByName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 50,
              "minLength": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "High Roller's Network Management"
        ],
        "operationId": "createPlayer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/highrollernetwork/player/{name}/referrer": {
      "get": {
        "tags": [
          "High Roller's Network Management"
        ],
        "operationId": "getPlayerReferrer",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 50,
              "minLength": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/highrollernetwork/player/{name}/downline": {
      "get": {
        "tags": [
          "High Roller's Network Management"
        ],
        "operationId": "getPlayerDownline",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 50,
              "minLength": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/highrollernetwork/player/{id}": {
      "get": {
        "tags": [
          "High Roller's Network Management"
        ],
        "operationId": "getPlayerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PlayerTransferRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1
          }
        },
        "required": [
          "name"
        ]
      },
      "PlayerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "referralChain": {
            "type": "string"
          },
          "exit": {
            "type": "boolean"
          }
        }
      },
      "PlayerRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1
          },
          "parentName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1
          }
        },
        "required": [
          "name",
          "parentName"
        ]
      }
    }
  }
}